% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_missingness.r
\name{predict_missingness_dv}
\alias{predict_missingness_dv}
\title{This function takes in events data, and a vector of labels for whether that value would be imputed or missing, and then fits
A model predicting that missingness
Tries to predict in terms of properties of each event, e.g. DV}
\usage{
predict_missingness_dv(label, print_every_n = 20)
}
\description{
This function takes in events data, and a vector of labels for whether that value would be imputed or missing, and then fits
A model predicting that missingness
Tries to predict in terms of properties of each event, e.g. DV
}
