% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_missingness.r
\name{predict_missingness_rhs}
\alias{predict_missingness_rhs}
\title{This function takes in events data, and a vector of labels for whether that value would be imputed or missing, and then fits
A model predicting that missingness
Tries to predict as a function of RHS covariates}
\usage{
predict_missingness_rhs(condition)
}
\description{
This function takes in events data, and a vector of labels for whether that value would be imputed or missing, and then fits
A model predicting that missingness
Tries to predict as a function of RHS covariates
}
